Change application and Deskbar replicant icon
Using App_Beam_2_NoMail from the Haiku tree and adding a
letter overlay to make it less bland. A variation of it is
used as no/mail indicator for the Deskbar replicant.

I removed the AFAIK unneeded declarations for the various
text/x-email bitmap icons from the rdef files. Those come
already as vector icons with Haiku.

Note: To get a "Archived Bitmap" from an icon:
* export to PNG
* drag into QuickRes
* double-click to edit open in QuickRes' icon editor
* select format "Archived Bitmap"
* save as rdef